+++ soltools/mkdepend/ifparser.h
@@ -65,8 +65,8 @@ typedef int Bool;
 typedef struct _if_parser {
     struct {                /* functions */
       const char *(*handle_error)  (struct _if_parser *, const char *, const char *);
-      int         (*eval_variable) (struct _if_parser *, const char *, int);
-      int         (*eval_defined)  (struct _if_parser *, const char *, int);
+      int         (*eval_variable) (struct _if_parser *, const char *, size_t);
+      int         (*eval_defined)  (struct _if_parser *, const char *, size_t);
     } funcs;
     char *data;
 } IfParser;
